16 спринт (последнее исправление)
